Roles done))))))

+++ Kursach/Controllers/AdminController.cs
@@ -5,9 +5,11 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Kursach.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Kursach.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class AdminController : Controller
     {
         RoleManager<IdentityRole> roleManager;
@@ -29,18 +31,20 @@ namespace Kursach.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(string SelectedCheckbox)
         {
-            System.Security.Claims.ClaimsPrincipal currentUser = this.User;
-            var currentId = userManager.GetUserId(currentUser);
+            var currentId = userManager.GetUserId(this.User);
             var SplitedValues = SelectedCheckbox.ToString().Split(',');
             foreach (string id in SplitedValues)
             {
                 User user = await userManager.FindByIdAsync(id);
                 if (user != null)
                 {
-                    await userManager.DeleteAsync(user);
-                    if (currentId == id)
+                    if (userManager.GetRolesAsync(user).Result.Count(x => x == "Admin") != 1)
                     {
-                        await signInManager.SignOutAsync();
+                        await userManager.DeleteAsync(user);
+                        if (currentId == id)
+                        {
+                            await signInManager.SignOutAsync();
+                        }
                     }
                 }
             }
@@ -59,12 +63,16 @@ namespace Kursach.Controllers
                 User user = await userManager.FindByIdAsync(id);
                 if (user != null)
                 {
-                    user.Status = StatusType.Blocked;
-                    await userManager.UpdateAsync(user);
-                    if (currentId == id)
+                    if (userManager.GetRolesAsync(user).Result.Count(x => x == "Admin") != 1)
                     {
-                        await signInManager.SignOutAsync();
+                        user.Status = StatusType.Blocked;
+                        await userManager.UpdateAsync(user);
+                        if (currentId == id)
+                        {
+                            await signInManager.SignOutAsync();
+                        }
                     }
+                    
                 }
             }
 
@@ -88,87 +96,37 @@ namespace Kursach.Controllers
             return RedirectToAction("index", "Admin", userManager.Users);
         }
 
-        //public IActionResult Create() => View();
-        //[HttpPost]
-        //public async Task<IActionResult> Create(string name)
-        //{
-        //    if (!string.IsNullOrEmpty(name))
-        //    {
-        //        IdentityResult result = await roleManager.CreateAsync(new IdentityRole(name));
-        //        if (result.Succeeded)
-        //        {
-        //            return RedirectToAction("Index");
-        //        }
-        //        else
-        //        {
-        //            foreach (var error in result.Errors)
-        //            {
-        //                ModelState.AddModelError(string.Empty, error.Description);
-        //            }
-        //        }
-        //    }
-        //    return View(name);
-        //}
+        [HttpPost]
+        public async Task<IActionResult> AddToAdmin(string SelectedCheckbox)
+        {
+            var SplitedValues = SelectedCheckbox.ToString().Split(',');
+            foreach (string id in SplitedValues)
+            {
+                User user = await userManager.FindByIdAsync(id);
+                if (user != null)
+                {
+                    await userManager.AddToRoleAsync(user, "Admin");
+                }
+            }
 
-        //[HttpPost]
-        //public async Task<IActionResult> Delete(string id)
-        //{
-        //    IdentityRole role = await roleManager.FindByIdAsync(id);
-        //    if (role != null)
-        //    {
-        //        IdentityResult result = await roleManager.DeleteAsync(role);
-        //    }
-        //    return RedirectToAction("Index");
-        //}
+            return RedirectToAction("index", "Admin", userManager.Users);
+        }
 
-        //public IActionResult UserList() => View(userManager.Users.ToList());
+        [HttpPost]
+        public async Task<IActionResult> RemoveFromAdmin(string SelectedCheckbox)
+        {
+            var SplitedValues = SelectedCheckbox.ToString().Split(',');
+            foreach (string id in SplitedValues)
+            {
+                User user = await userManager.FindByIdAsync(id);
+                if (user != null)
+                {
+                    await userManager.RemoveFromRoleAsync(user, "Admin");
+                }
+            }
 
-        //public async Task<IActionResult> Edit(string userId)
-        //{
-        //    // получаем пользователя
-        //    User user = await userManager.FindByIdAsync(userId);
-        //    if (user != null)
-        //    {
-        //        // получем список ролей пользователя
-        //        var userRoles = await userManager.GetRolesAsync(user);
-        //        var allRoles = roleManager.Roles.ToList();
-        //        ChangeRoleModel model = new ChangeRoleModel
-        //        {
-        //            UserId = user.Id,
-        //            UserEmail = user.Email,
-        //            UserRoles = userRoles,
-        //            AllRoles = allRoles
-        //        };
-        //        return View(model);
-        //    }
-
-        //    return NotFound();
-        //}
-        //[HttpPost]
-        //public async Task<IActionResult> Edit(string userId, List<string> roles)
-        //{
-        //    // получаем пользователя
-        //    User user = await userManager.FindByIdAsync(userId);
-        //    if (user != null)
-        //    {
-        //        // получем список ролей пользователя
-        //        var userRoles = await userManager.GetRolesAsync(user);
-        //        // получаем все роли
-        //        var allRoles = roleManager.Roles.ToList();
-        //        // получаем список ролей, которые были добавлены
-        //        var addedRoles = roles.Except(userRoles);
-        //        // получаем роли, которые были удалены
-        //        var removedRoles = userRoles.Except(roles);
-
-        //        await userManager.AddToRolesAsync(user, addedRoles);
-
-        //        await userManager.RemoveFromRolesAsync(user, removedRoles);
-
-        //        return RedirectToAction("UserList");
-        //    }
-
-        //    return NotFound();
-        //}
+            return RedirectToAction("index", "Admin", userManager.Users);
+        }
     }
 }
